Unit test for single stage
Summary:
In this test we ask the solver to create one single stage with all the operations of the model.
In practice we would allow the solver to split the model into multiple stages
but here we want to check that the memory analysis done by the solver is similar to the actual
memory used by the model and, in particular, that the solver is not missing any operation or
variable.

Reviewers: #analysis_tools_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, simonl, Richardc, davidt

Reviewed By: #analysis_tools_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Richardc

Maniphest Tasks: T56694

Differential Revision: https://phabricator.sourcevertex.net/D61695